Phase 1: Attendance/Leave (Missing functionalities) 1.1 Fix LeaveApplication.NumberOfDays type (int → decimal) 1.2 Face-match threshold enforcement 1.3 Check-Out time capture 1.4 Holiday calendar entity + CRUD + UI 1.5 Payroll attendance summary page

+++ src/CmScheme.AttendanceLeave.Application/Features/Leave/ApplyLeave/ApplyLeaveCommand.cs
@@ -9,7 +9,7 @@ public sealed record ApplyLeaveCommand : ICommand<Result<int>>
     public string LeaveType { get; init; } = string.Empty;
     public DateTime FromDate { get; init; }
     public DateTime ToDate { get; init; }
-    public int NumberOfDays { get; init; }
+    public decimal NumberOfDays { get; init; }
     public string HalfDayFullDay { get; init; } = string.Empty;
     public string LeaveReason { get; init; } = string.Empty;
     public string? AttachmentPath { get; init; }
